requirements.txt
Flask==3.1.2
fpdf2
gunicorn
psycopg2-binary
supabase
